requirements.txt
cffi==1.9.1
coverage==4.3.1
coveralls==1.1
cryptography==1.7.1
dnspython==1.15.0
docopt==0.6.2
docutils==0.13.1
Flask==0.10.1
Flask-Cors==3.0.2
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.1
httpretty==0.8.14
idna==2.2
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.9.0
logstash-formatter==0.5.16
MarkupSafe==0.23
mock==2.0.0
mysqlclient==1.3.9
pbr==1.10.0
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
python-consul==0.7.0
python-dateutil==2.6.0
python-magic==0.4.6
pytz==2016.10
PyYAML==3.12
requests==2.7.0
requests-futures==0.9.5
s3transfer==0.1.10
six==1.10.0
SQLAlchemy==1.1.5
tornado==4.4.2
Werkzeug==0.10.1
